# backend/internal/http/http.go
// Package http 是 erp-inventory 的 REST 面（对外路径前缀 /erp/inventory，
// 与 assembly.yaml 的 edge_routes 一致），契约是 contracts/inventory.openapi.yaml。
// ⚠️ TCC 四件套（Reserve/CancelReservation/ConfirmIssue/GetReservationStatus）
// 与 BatchGetBalance 永远不进 REST：它们是组件间协议，不是人类操作。
package http

import (
	"fmt"
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"

	besdk "github.com/brickKit/be-sdk-go"
	"github.com/brickKit/erp-inventory/v2/backend/internal/repo"
	"github.com/brickKit/erp-inventory/v2/backend/internal/service"
)

// RegisterRoutes 挂载业务路由，每条都带 assembly.yaml 里声明的权限键。
//
// ⚠️ `erp.inventory.reserve` / `.issue` 两个已声明的权限键在这里用不上——
// Reserve / ConfirmIssue 永远不进 REST（见包文档），它们只是权限目录里的
// 条目，供角色配置与审计展示，不对应任何 besdk.GET/POST 调用点，不是漏填。
// warehouse 维数据范围（谁能看哪个仓库）不在 JWT 里：service 层用
// besdk.ScopeOf 取 sub，再查本组件自己的 warehouse_access 表。
func RegisterRoutes(eng *gin.Engine, svc *service.Service) {
	g := eng.Group("/erp/inventory")
	besdk.POST(g, "/movements/receive", "erp.inventory.receive", receiveHandler(svc))
	besdk.POST(g, "/movements/adjust", "erp.inventory.adjust", adjustHandler(svc))
	besdk.GET(g, "/balances", "erp.inventory.view", getBalanceHandler(svc))
	besdk.GET(g, "/balances/list", "erp.inventory.view", listBalancesHandler(svc))
	besdk.GET(g, "/warehouses", "erp.inventory.view", listWarehousesHandler(svc))
	besdk.GET(g, "/stats/summary", "erp.inventory.view", statsSummaryHandler(svc))
	besdk.GET(g, "/movements", "erp.inventory.view", listMovementsHandler(svc))
	besdk.GET(g, "/warehouse-access/:sub", "erp.inventory.manage_access", listWarehouseAccessHandler(svc))
	besdk.POST(g, "/warehouse-access/:sub", "erp.inventory.manage_access", grantWarehouseAccessHandler(svc))
	besdk.DELETE(g, "/warehouse-access/:sub/:warehouse_id", "erp.inventory.manage_access", revokeWarehouseAccessHandler(svc))
}

type receiveRequest struct {
	IdempotencyKey string `json:"idempotency_key" binding:"required"`
	ProductID      string `json:"product_id" binding:"required"`
	WarehouseID    string `json:"warehouse_id" binding:"required"`
	Qty            string `json:"qty" binding:"required"`
	BatchNo        string `json:"batch_no"`
	SerialNo       string `json:"serial_no"`
}

func receiveHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req receiveRequest
		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		movementID, err := svc.Receive(c.Request.Context(), repo.ReceiveInput{
			IdempotencyKey: req.IdempotencyKey, ProductID: req.ProductID, WarehouseID: req.WarehouseID,
			Qty: req.Qty, BatchNo: req.BatchNo, SerialNo: req.SerialNo,
		})
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.JSON(http.StatusOK, gin.H{"movement_id": movementID})
	}
}

type adjustRequest struct {
	IdempotencyKey string `json:"idempotency_key" binding:"required"`
	ProductID      string `json:"product_id" binding:"required"`
	WarehouseID    string `json:"warehouse_id" binding:"required"`
	QtyDelta       string `json:"qty_delta" binding:"required"`
	Reason         string `json:"reason"`
}

func adjustHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req adjustRequest
		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		movementID, err := svc.Adjust(c.Request.Context(), repo.AdjustInput{
			IdempotencyKey: req.IdempotencyKey, ProductID: req.ProductID, WarehouseID: req.WarehouseID,
			QtyDelta: req.QtyDelta, Reason: req.Reason,
		})
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.JSON(http.StatusOK, gin.H{"movement_id": movementID})
	}
}

func toBalanceDTO(b *repo.Balance) gin.H {
	return gin.H{
		"product_id": b.ProductID, "warehouse_id": b.WarehouseID,
		"on_hand_qty": b.OnHandQty, "reserved_qty": b.ReservedQty, "available_qty": b.AvailableQty,
		"version": b.Version,
	}
}

func getBalanceHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		b, err := svc.GetBalance(c.Request.Context(), c.Query("product_id"), c.Query("warehouse_id"))
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.JSON(http.StatusOK, toBalanceDTO(b))
	}
}

const rfc3339 = "2006-01-02T15:04:05.999999999Z07:00"

func toMovementDTO(m *repo.Movement) gin.H {
	return gin.H{
		"id": m.ID, "product_id": m.ProductID, "warehouse_id": m.WarehouseID, "qty": m.Qty,
		"reason": m.Reason, "order_id": m.OrderID, "batch_no": m.BatchNo, "serial_no": m.SerialNo,
		"created_at": m.CreatedAt.Format(rfc3339),
	}
}

// parseListMovementsInput 读 GET /movements 的查询参数。created_after /
// created_before 是 RFC 3339 时间，不传就由 repo 补默认窗口；格式不对返回错误
// （handler 回 400），不悄悄忽略。
func parseListMovementsInput(c *gin.Context) (repo.ListInput, error) {
	pageSize, _ := strconv.Atoi(c.Query("page_size"))
	in := repo.ListInput{
		Cursor: c.Query("cursor"), PageSize: pageSize,
		ProductID: c.Query("product_id"), WarehouseID: c.Query("warehouse_id"),
	}
	for _, p := range []struct {
		name string
		dst  *time.Time
	}{{"created_after", &in.CreatedAfter}, {"created_before", &in.CreatedBefore}} {
		raw := c.Query(p.name)
		if raw == "" {
			continue
		}
		t, err := time.Parse(time.RFC3339, raw)
		if err != nil {
			return in, fmt.Errorf("%s 不是 RFC 3339 时间：%q", p.name, raw)
		}
		*p.dst = t
	}
	return in, nil
}

func listMovementsHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		in, err := parseListMovementsInput(c)
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		out, err := svc.ListMovements(c.Request.Context(), in)
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		dtos := make([]gin.H, 0, len(out.Movements))
		for _, m := range out.Movements {
			dtos = append(dtos, toMovementDTO(m))
		}
		c.JSON(http.StatusOK, gin.H{"movements": dtos, "next_cursor": out.NextCursor})
	}
}

// ── warehouse_access 管理（erp.inventory.manage_access）──

func listWarehouseAccessHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		ids, err := svc.ListWarehouseAccess(c.Request.Context(), c.Param("sub"))
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		// REST 契约里 id 一律是字符串（同 warehouse_id 在 Balance/Movement
		// DTO 里的既有约定）。
		out := make([]string, len(ids))
		for i, id := range ids {
			out[i] = strconv.FormatInt(id, 10)
		}
		c.JSON(http.StatusOK, gin.H{"warehouse_ids": out})
	}
}

type grantWarehouseAccessRequest struct {
	WarehouseID string `json:"warehouse_id" binding:"required"`
}

func grantWarehouseAccessHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req grantWarehouseAccessRequest
		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		if err := svc.GrantWarehouseAccess(c.Request.Context(), c.Param("sub"), req.WarehouseID); err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.Status(http.StatusOK)
	}
}

func revokeWarehouseAccessHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		err := svc.RevokeWarehouseAccess(c.Request.Context(), c.Param("sub"), c.Param("warehouse_id"))
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		c.Status(http.StatusOK)
	}
}

// ── 仓库、余额列表、库存统计（都按调用者的 warehouse_access 过滤）──

func listWarehousesHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		whs, err := svc.ListWarehouses(c.Request.Context())
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		out := make([]gin.H, 0, len(whs))
		for _, w := range whs {
			out = append(out, gin.H{"id": w.ID, "code": w.Code, "name": w.Name, "status": w.Status})
		}
		c.JSON(http.StatusOK, gin.H{"warehouses": out})
	}
}

func listBalancesHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		pageSize, _ := strconv.Atoi(c.Query("page_size"))
		out, err := svc.ListBalances(c.Request.Context(), repo.BalanceListInput{
			Cursor: c.Query("cursor"), PageSize: pageSize,
			ProductID: c.Query("product_id"), WarehouseID: c.Query("warehouse_id"),
		})
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		dtos := make([]gin.H, 0, len(out.Balances))
		for _, b := range out.Balances {
			dtos = append(dtos, toBalanceDTO(b))
		}
		c.JSON(http.StatusOK, gin.H{"balances": dtos, "next_cursor": out.NextCursor})
	}
}

func statsSummaryHandler(svc *service.Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		sum, err := svc.StockSummary(c.Request.Context())
		if err != nil {
			_ = c.Error(service.ToStatus(err))
			return
		}
		low := make([]gin.H, 0, len(sum.LowStock))
		for _, it := range sum.LowStock {
			low = append(low, gin.H{
				"product_id": it.ProductID, "warehouse_id": it.WarehouseID, "available_qty": it.AvailableQty,
			})
		}
		c.JSON(http.StatusOK, gin.H{
			"sku_count": sum.SKUCount, "total_on_hand_qty": sum.TotalOnHandQty,
			"low_stock": low, "low_stock_count": sum.LowStockCount,
		})
	}
}