# contracts/inventory.openapi.yaml
openapi: 3.0.3
info:
  title: erp-inventory · 库存管理
  version: 2.0.0
  description: |
    对外 REST 面；别的组件走 gRPC（contracts/erp/inventory/v1/inventory.proto）。
    ⚠️ 不暴露 Reserve/CancelReservation/ConfirmIssue/GetReservationStatus——
    这四个是组件间 TCC 协议的一部分，不是人类操作。
    人类操作是"出库单"，那在 erp-sales / erp-purchase 侧。
    BatchGetBalance 不暴露：它是给其他组件 gRPC 客户端防 N+1 用的批量读。
    余额、流水、仓库、统计与入库 / 调整都只认调用者 warehouse_access 里授权的仓库；
    /warehouse-access 是给管理员分配这份授权用的。

servers:
  - url: /erp/inventory

# ⚠️ /healthz 不在这份契约里：它是 be-sdk-go 的 NewGinEngine 给每个组件统一
# 挂的裸根路径，不带 /erp/inventory 前缀。

paths:
  /movements/receive:
    post:
      summary: Receive（入库）
      operationId: receiveInventory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReceiveRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceiveResponse" }

  /movements/adjust:
    post:
      summary: Adjust（盘盈盘亏）
      operationId: adjustInventory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdjustRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdjustResponse" }

  /balances:
    get:
      summary: GetBalance
      operationId: getInventoryBalance
      parameters:
        - name: product_id
          in: query
          required: true
          schema: { type: string }
        - name: warehouse_id
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Balance" }

  /balances/list:
    get:
      summary: ListBalances（余额列表，按调用者有授权的仓库过滤）
      operationId: listInventoryBalances
      description: |
        余额是当前状态，行数上限是「产品数 × 仓库数」，所以没有默认时间窗口；
        按余额行 id 升序、游标分页。`GET /balances` 仍是单条查询，语义不变。
      parameters:
        - name: product_id
          in: query
          schema: { type: string }
        - name: warehouse_id
          in: query
          schema: { type: string }
          description: 调用者没有这个仓库的授权时结果为空
        - name: cursor
          in: query
          schema: { type: string }
          description: 游标分页，无 offset 参数；非法游标返回 400
        - name: page_size
          in: query
          schema: { type: integer }
          description: 默认 50，上限 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BalanceList" }

  /warehouses:
    get:
      summary: ListWarehouses（调用者有授权的仓库）
      operationId: listInventoryWarehouses
      description: |
        仓库是迁移播种的主数据（几十行量级），一次返回调用者有授权的全部仓库，
        按 id 升序，不分页。没有任何授权时返回空列表。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WarehouseList" }

  /stats/summary:
    get:
      summary: StockSummary（仪表盘库存统计，按调用者有授权的仓库统计）
      operationId: getInventoryStatsSummary
      description: |
        sku_count：可见仓库里在手量大于 0 的不同产品数。total_on_hand_qty：可见
        仓库的在手量合计。low_stock：可用量（在手 - 已预留）低于配置项
        LOW_STOCK_THRESHOLD 的余额行，按可用量升序，最多 50 条；low_stock_count
        是满足条件的总行数。比较全程用十进制，不经过浮点数。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockSummary" }

  /movements:
    get:
      summary: ListMovements
      operationId: listInventoryMovements
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
          description: 游标分页，无 offset 参数——深分页在契约层面不可表达；非法游标返回 400
        - name: page_size
          in: query
          schema: { type: integer }
        - name: product_id
          in: query
          schema: { type: string }
        - name: warehouse_id
          in: query
          schema: { type: string }
        - name: created_after
          in: query
          schema: { type: string, format: date-time }
        - name: created_before
          in: query
          schema: { type: string, format: date-time }
          description: RFC 3339。不传时间范围时默认最近 90 天；格式不对返回 400
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  movements:
                    type: array
                    items: { $ref: "#/components/schemas/Movement" }
                  next_cursor: { type: string }

components:
  schemas:
    MovementReason:
      type: string
      enum: [RECEIVE, ISSUE, ADJUST_GAIN, ADJUST_LOSS]

    Balance:
      type: object
      properties:
        product_id: { type: string }
        warehouse_id: { type: string }
        on_hand_qty:
          type: string
          description: 金额/数量一律用 string 传 decimal，绝不用 double（跨语言精度）
        reserved_qty: { type: string }
        available_qty:
          type: string
          description: on_hand_qty - reserved_qty，服务端算好直接给
        version: { type: integer, format: int64 }

    BalanceList:
      type: object
      properties:
        balances:
          type: array
          items: { $ref: "#/components/schemas/Balance" }
        next_cursor: { type: string, description: 空串表示没有下一页 }

    Warehouse:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        status: { type: string }

    WarehouseList:
      type: object
      properties:
        warehouses:
          type: array
          items: { $ref: "#/components/schemas/Warehouse" }

    LowStockItem:
      type: object
      properties:
        product_id: { type: string }
        warehouse_id: { type: string }
        available_qty: { type: string, description: decimal-as-string }

    StockSummary:
      type: object
      properties:
        sku_count: { type: integer, format: int64 }
        total_on_hand_qty: { type: string, description: decimal-as-string }
        low_stock:
          type: array
          items: { $ref: "#/components/schemas/LowStockItem" }
        low_stock_count: { type: integer, format: int64 }

    Movement:
      type: object
      properties:
        id: { type: string }
        product_id: { type: string }
        warehouse_id: { type: string }
        qty: { type: string, description: 带符号，decimal-as-string }
        reason: { $ref: "#/components/schemas/MovementReason" }
        order_id: { type: string }
        batch_no: { type: string }
        serial_no: { type: string }
        created_at: { type: string, format: date-time }

    ReceiveRequest:
      type: object
      required: [idempotency_key, product_id, warehouse_id, qty]
      properties:
        idempotency_key:
          type: string
          description: 跨重试幂等键，被调用方先原子声明再执行，重复请求返回第一次的结果
        product_id: { type: string }
        warehouse_id: { type: string }
        qty: { type: string }
        batch_no: { type: string }
        serial_no: { type: string }

    ReceiveResponse:
      type: object
      properties:
        movement_id: { type: string }

    AdjustRequest:
      type: object
      required: [idempotency_key, product_id, warehouse_id, qty_delta, reason]
      properties:
        idempotency_key: { type: string }
        product_id: { type: string }
        warehouse_id: { type: string }
        qty_delta:
          type: string
          description: 带符号：正为盘盈，负为盘亏
        reason:
          type: string
          description: 人类可读的调整原因（落进流水的 note）；调整不走审批

    AdjustResponse:
      type: object
      properties:
        movement_id: { type: string }